+++ src/org/lwjgl/demo/opengl/raytracing/HybridDemo.java
@@ -554,7 +554,7 @@ public class HybridDemo {
 		 * compute the blend factor between old frame and new frame. 0.0 - use
 		 * only the new frame > 0.0 - blend between old frame and new frame
 		 */
-		float blendFactor = (float) frameNumber / ((float) frameNumber + 1.0f);
+		float blendFactor = frameNumber / (frameNumber + 1.0f);
 		glUniform1f(blendFactorUniform, blendFactor);
 		glUniform1i(bounceCountUniform, bounceCount);
